Updated README.md - optimize console command's
The previous console commands, made it impossible to use the GUI copy function as intended:

If someone clicked on the "copy" button to the right of the command, then not only the command was copied, but also the preceding "$" character.

The unnecessary character was removed from all commands and the usual Markdown tag "console" was added.